fastapi-auth/requirements/base.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
slowapi>=0.1.9
email-validator>=2.1.0
